Changed name of the mod + Commited Flags
I changed the name of the mod since it's not a portrait mod anymore. I also added the flags and backgrounds that I had as a separate mod from the beginning. Though these are Arumba specific and it will probably be better if they only get included in the "Arumba as a leader" mod